Web/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv>=1.0.1
pydantic>=2.6.4
cryptography>=42.0.8
pycryptodome>=3.20.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
bcrypt>=4.1.0
starlette[cors]